+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this provider will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### :rocket: Added
+* Added Route53 Domains resource
+
 ## [v0.6.2] - 2021-11-03
 ###### SDK Version: 0.5.1